Force dump when dirs are modified
When files were modified and added in the same event
loadFiles() skipped the dump and called loadAllFiles()
to check for new files. If the new files were empty
loadAllFiles() wouldn't dump since from its point of view,
nothing changed.

The fix adds a param to loadAllFiles() to acknowledge that
something has indeed changed.