SERVER-135 Create a new zval before casting it to string to avoid side effects
The convert_to_string call change the type of the zval and this can
be problematic. This zval can be used anywhere in the code and if we
change the type, we creating side effects on a parameter that should be
only used by value. We need to create a new zval in order to make this
casting safe.